PS-109 Filters completed in chip format, filters are now stackable, search terms can be added via add filter button and are persistent, custom filters can be deleted, and ui cleaned up.

+++ gateway/src/main/java/com/patreonshout/beans/request/FilterDeleteRequest.java
@@ -17,5 +17,5 @@ public class FilterDeleteRequest extends FilterCommonRequest implements Serializ
      */
     @NotNull
     @JsonProperty("filter_id")
-    private int filterId;
+    private long filterId;
 }
